+++ python/ray/air/_internal/util.py
@@ -2,9 +2,7 @@ import copy
 import logging
 import os
 import queue
-import socket
 import threading
-from contextlib import closing
 from typing import Optional
 
 import numpy as np
@@ -12,13 +10,6 @@ import numpy as np
 from ray.air.constants import _ERROR_REPORT_TIMEOUT
 
 logger = logging.getLogger(__name__)
-
-
-def find_free_port(family=socket.AF_INET):
-    with closing(socket.socket(family, socket.SOCK_STREAM)) as s:
-        s.bind(("", 0))
-        s.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-        return s.getsockname()[1]
 
 
 def is_nan(value):
