+++ python/ray/tests/conftest.py
@@ -22,14 +22,13 @@ import pytest
 
 import ray
 import ray._private.ray_constants as ray_constants
-from ray._common.network_utils import build_address
+from ray._common.network_utils import build_address, find_free_port
 from ray._common.test_utils import wait_for_condition
 from ray._private.conftest_utils import set_override_dashboard_url  # noqa: F401
 from ray._private.runtime_env import virtualenv_utils
 from ray._private.test_utils import (
     RayletKiller,
     external_redis_test_enabled,
-    find_free_port,
     get_and_run_resource_killer,
     get_redis_cli,
     init_error_pubsub,
