+++ python/ray/tests/test_gcs_utils.py
@@ -11,11 +11,10 @@ import redis
 import ray
 import ray._private.gcs_utils as gcs_utils
 import ray._private.ray_constants as ray_constants
-from ray._common.network_utils import parse_address
+from ray._common.network_utils import find_free_port, parse_address
 from ray._common.test_utils import async_wait_for_condition
 from ray._private.test_utils import (
     external_redis_test_enabled,
-    find_free_port,
     generate_system_config_map,
 )
 from ray._raylet import GcsClient, NodeID
