+++ python/ray/tests/test_ray_event_export_task_events.py
@@ -8,9 +8,10 @@ import pytest
 
 import ray
 import ray.dashboard.consts as dashboard_consts
+from ray._common.network_utils import find_free_port
 from ray._common.test_utils import wait_for_condition
 from ray._private import ray_constants
-from ray._private.test_utils import find_free_port, run_string_as_driver_nonblocking
+from ray._private.test_utils import run_string_as_driver_nonblocking
 from ray._raylet import GcsClient
 
 logger = logging.getLogger(__name__)
